requirements.txt
Django==5.2.7
django-allauth==65.3.0
django-bootstrap5==24.3
Pillow==11.0.0